update models to content-types (#5)
