+++ utils/strapi-server.js
@@ -1,1 +1,1 @@
-module.exports = require('./src/server')
+module.exports = require('./server')
